Title: 앞의 두 수를 더한 수열 2, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231120/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.cpp
@@ -2,11 +2,11 @@
 
 using namespace std;
 int main() {
-    int a = 0;
-    int b = 1;
+    long a = 0;
+    long b = 1;
 
     int n;
-    int num = 0;
+    long num = 0;
     cin >> n;
 
     for(int i = 2; i <= n; i++){
